crons/top100_table/requirements.txt
psycopg2==2.9.10
requests==2.32.3
python-dotenv==1.0.1